Refactoring following correcting the spelling mistake for the name of the  'exception' package (from 'exeption')
