+++ simulator/src/main/java/kcl/teamIndexZero/traffic/simulator/SimulatorEntryPoint.java
@@ -9,8 +9,8 @@ import kcl.teamIndexZero.traffic.simulator.data.SimulationParams;
 import kcl.teamIndexZero.traffic.simulator.data.descriptors.LinkDescription;
 import kcl.teamIndexZero.traffic.simulator.data.descriptors.RoadDescription;
 import kcl.teamIndexZero.traffic.simulator.data.features.Junction;
-import kcl.teamIndexZero.traffic.simulator.exeptions.AlreadyExistsException;
-import kcl.teamIndexZero.traffic.simulator.exeptions.MapIntegrityException;
+import kcl.teamIndexZero.traffic.simulator.exceptions.AlreadyExistsException;
+import kcl.teamIndexZero.traffic.simulator.exceptions.MapIntegrityException;
 import kcl.teamIndexZero.traffic.simulator.mapSetup.MapFactory;
 import kcl.teamIndexZero.traffic.simulator.mapSetup.mapFeatureType;
