+++ simulator/src/main/java/kcl/teamIndexZero/traffic/simulator/SimulatorFactory.java
@@ -5,7 +5,7 @@ package kcl.teamIndexZero.traffic.simulator;
  */
 public interface SimulatorFactory {
     /**
-     * Create a new, or get an existing one, simulator.
+     * Create a new, or get an existing in, simulator.
      *
      * @return simulator instance.
      */
