+++ simulator/src/main/java/kcl/teamIndexZero/traffic/simulator/data/SimulationTick.java
@@ -8,7 +8,7 @@ import java.time.format.DateTimeFormatter;
  * For now we are mostly intersted in several things:
  * <p>
  * 1. What's the ordinal number of this tick?
- * 2. How long does this tick span for (in simulated time)? E.g. one tick may be equal 1 second or 1 minute, that is
+ * 2. How long does this tick span for (in simulated time)? E.g. in tick may be equal 1 second or 1 minute, that is
  * our simulation granularity
  * 3. What is the tick's simulated moment in time? (it may be used in future - consider, for example, night time
  * approaching or dusk forcing drivers to lower the speed so that they can react, or a time dependency for traffic
