+++ simulator/src/main/java/kcl/teamIndexZero/traffic/simulator/exceptions/MapIntegrityException.java
@@ -1,4 +1,4 @@
-package kcl.teamIndexZero.traffic.simulator.exeptions;
+package kcl.teamIndexZero.traffic.simulator.exceptions;
 
 /**
  * Created by Es on 28/02/2016.
